Refactor to use a service
