+++ package.json
@@ -4,7 +4,7 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "mocha tests --recursive --watch"
   },
   "author": "",
   "license": "ISC",
@@ -12,6 +12,7 @@
     "express": "^4.14.0"
   },
   "devDependencies": {
+    "chai": "^3.5.0",
     "nodemon": "^1.11.0"
   }
 }
